Release v0.15.0



Merged-by: Guillermo Marcus <4169784+gmarcusm@users.noreply.github.com>

+++ sionna/ofdm/pilot_pattern.py
@@ -254,7 +254,8 @@ class EmptyPilotPattern(PilotPattern):
                       num_effective_subcarriers]
         mask = tf.zeros(shape, tf.bool)
         pilots = tf.zeros(shape[:2]+[0], dtype)
-        super().__init__(mask, pilots, trainable=False, normalize=False)
+        super().__init__(mask, pilots, trainable=False, normalize=False,
+                         dtype=dtype)
 
 class KroneckerPilotPattern(PilotPattern):
     """Simple orthogonal pilot pattern with Kronecker structure.
